@@ -4,31 +4,31 @@
   "description": "Find issues to work on with the labels and keywords you specify.",
   "private": true,
   "dependencies": {
-    "@octokit/rest": "^16.28.2",
-    "axios": "^0.21.1",
+    "@octokit/rest": "^19.0.3",
+    "axios": "^0.27.2",
     "chromatism": "^3.0.0",
-    "dotenv-webpack": "^7.1.1",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.26",
+    "dotenv": "^16.0.1",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.34",
     "netlify-lambda": "^2.0.15",
     "npm-run-all": "^4.1.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-ga4": "^1.4.1",
-    "react-mde": "^10.0.3",
-    "react-moment": "^0.9.2",
+    "react-markdown": "^8.0.3",
+    "react-moment": "^1.1.2",
     "react-router-dom": "^6.3.0",
     "react-toggle-switch": "^3.0.4",
-    "showdown": "^1.9.1"
+    "showdown": "^2.1.0"
   },
   "devDependencies": {
-    "@testing-library/react": "^11.0.4",
-    "eslint-plugin-react": "^7.20.6",
-    "eslint-plugin-react-hooks": "^4.1.0",
-    "mockdate": "^3.0.2",
-    "prettier": "^2.1.1",
+    "@testing-library/react": "^13.3.0",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "mockdate": "^3.0.5",
+    "prettier": "^2.7.1",
     "react-scripts": "^5.0.1",
-    "react-test-renderer": "^16.13.1"
+    "react-test-renderer": "^18.2.0"
   },
   "scripts": {
     "start": "run-p start:**",
@@ -37,7 +37,7 @@
     "build:app": "react-scripts build",
     "test:app": "react-scripts test --env=jsdom",
     "eject:app": "react-scripts eject",
-    "start:lambda": "netlify-lambda serve src/api --config ./webpack.functions.js",
+    "start:lambda": "netlify-lambda serve src/api",
     "build:lambda": "netlify-lambda build src/api",
     "lint": "eslint 'src/**/*.js'",
     "format-check": "prettier 'src/**/*.{js,css,json}' --check",
