@@ -1,6 +1,2 @@
-const Dotenv = require('dotenv-webpack');
-
 // @see https://github.com/netlify/netlify-lambda#webpack-configuration
-module.exports = {
-  plugins: [new Dotenv()],
-};
+module.exports = {};
